Restart url-fetcher & tweet-loader
These daemons need to run for the docnow application to be able to
collect tweets and display information about webpages referenced in
tweets. This commit modifies the systemd config so that they are
restarted if they exit for any reason.

I think this is a quick fix for https://github.com/DocNow/docnow/issues/173